+++ Board-code/HiL_Board/HiL_code/Inc/HiL_mcu_commands.h
@@ -17,11 +17,11 @@
 extern DAC_HandleTypeDef hdac;
 
 
-int HiL_mcu_commands_potentiometer_emulator(int value1, int value2);
+uint8_t HiL_mcu_commands_potentiometer_emulator(uint8_t value1, uint8_t value2);
 
-int HiL_mcu_commands_binary_action (int binary_object, int desired_state);
+uint8_t HiL_mcu_commands_binary_action (uint8_t binary_object, uint8_t desired_state);
 
-int HiL_mcu_commands_binary_status (int binary_object);
+uint8_t HiL_mcu_commands_binary_status (uint8_t binary_object);
 
 
 #endif /* USER_COMMANDS_H_ */
